refactor(bindings): move collateral-return to a folder surface
