+++ packages/bindings/package.json
@@ -23,8 +23,8 @@
       "default": "./dist/clob/index.js"
     },
     "./collateral-return": {
-      "types": "./dist/collateral-return.d.ts",
-      "default": "./dist/collateral-return.js"
+      "types": "./dist/collateral-return/index.d.ts",
+      "default": "./dist/collateral-return/index.js"
     },
     "./gamma": {
       "types": "./dist/gamma/index.d.ts",
@@ -57,7 +57,7 @@
         "./dist/clob/index.d.ts"
       ],
       "collateral-return": [
-        "./dist/collateral-return.d.ts"
+        "./dist/collateral-return/index.d.ts"
       ],
       "data": [
         "./dist/data/index.d.ts"
