+++ packages/bindings/src/collateral-return/plan.ts
@@ -1,16 +1,12 @@
 import { expectHexString, type HexString } from '@polymarket/types';
 import { z } from 'zod';
-import {
-  RelayerDepositWalletExecuteRequestSchema,
-  RelayerLegacyExecuteRequestSchema,
-} from './relayer';
-import type { DecimalString, EvmAddress, PositionId } from './shared';
+import type { DecimalString, EvmAddress, PositionId } from '../shared';
 import {
   DecimalStringSchema,
   E6BigIntStringToDecimalStringSchema,
   EvmAddressSchema,
   PositionIdSchema,
-} from './shared';
+} from '../shared';
 
 const HexStringSchema = z.string().transform((value) => expectHexString(value));
 
@@ -194,21 +190,3 @@ export const CollateralReturnPlanResponseSchema = z
       wallet: response.wallet,
     }),
   );
-
-/**
- * Submission payload for a planned collateral return: the plan identifier and
- * the signed relayer envelope carrying the plan's exact Router call. The
- * envelope is a Deposit Wallet batch request or a legacy Safe/Proxy relay
- * request.
- */
-export const CollateralReturnSubmitRequestSchema = z.object({
-  plan_hash: HexStringSchema,
-  envelope: z.union([
-    RelayerDepositWalletExecuteRequestSchema,
-    RelayerLegacyExecuteRequestSchema,
-  ]),
-});
-
-export type CollateralReturnSubmitRequest = z.input<
-  typeof CollateralReturnSubmitRequestSchema
->;
